kite-handler/terminal: make Terminal Handler a struct to modify certain things
